Clean up repeatable form, keys
