+++ src/hooks/use-student-spotlight-reducer.js
@@ -1,7 +1,7 @@
 import { useReducer } from 'react';
 import { studentSpotlightReducer } from '~utils/student-spotlight-reducer';
 
-const initialState = { students: [{}] };
+const initialState = { students: [{ key: 0 }] };
 
 export const useStudentSpotlightReducer = () => {
     const [state, dispatch] = useReducer(studentSpotlightReducer, initialState);
